release: use the latest tag to update version information (#95)
## Issue

Previously, we needed to wait for the release page creation to complete
before running `rake release:version:update`. However, that step was
sometimes overlooked, and the task was run before the release page
existed. As a result, this task would fail with an error when the
release page was unavailable.

## Solution

By switching to the latest tag instead of the latest release, we ensure
that the version update process no longer depends on the presence of the
release page. As soon as a tag is pushed, we can execute this rake task
to retrieve the release version.